Render graphtool with React
