+++ app/javascript/components/GraphTool/nested/Accessories.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function Accessories() {
+  return (
+    <div className="accessories"></div>
+  );
+}
+
+export default Accessories;
