+++ app/javascript/components/GraphTool/nested/EqOverlay.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function EqOverlay() {
+  return (
+    <div style={{ display: 'none' }} className="extra-eq-overlay">
+      AutoEQ is running, it could take 5~20 seconds or more.
+    </div>
+  );
+}
+
+export default EqOverlay;
