+++ app/javascript/components/GraphTool/nested/ExternalLinks.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function ExternalLinks() {
+  return (
+    <div className="external-links"></div>
+  );
+}
+
+export default ExternalLinks;
